pyproject.toml
[tool.poetry]
name = "mysite"
version = "0.1.0"
description = ""
authors = ["Isaac Wangombe <isaacwangombe5@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.7"
django = "^3.2.16"
dj-database-url = "^1.0.0"
psycopg2-binary = "^2.9.5"
gunicorn = "^20.1.0"
whitenoise = {extras = ["brotli"], version = "^6.2.0"}
djangorestframework = "^3.14.0"
django-cors-headers = "^3.13.0"
cloudinary = "^1.30.0"
django-bootstrap5 = "^22.2"


[build-system]
requires = ["poetry-core",
"setuptools!=50.0"]
build-backend = "poetry.core.masonry.api"
